Rename App interactor to Page
See #582

While the name `App` works pretty well for `visit`, we would like to add
further functionality to this interactor, like checking global
attributes and in that case the name works decidedly less well and
`Page` sounds better.
