+++ .changeset/lucky-gorillas-work.md
@@ -1,0 +1,6 @@
+---
+"@bigtest/interactor": minor
+"bigtest": minor
+---
+
+Rename App interactor to Page and deprecate App interactor
