+++ packages/interactor/src/index.ts
@@ -1,5 +1,6 @@
 export { Interactor } from './interactor';
 export { createInteractor } from './create-interactor';
+export { Page } from './page';
 export { App } from './app';
 export { perform } from './perform';
 export { fillIn } from './fill-in';
