+++ packages/interactor/test/helpers.ts
@@ -7,6 +7,7 @@ let jsdom: JSDOM;
 export function dom(html: string) {
   jsdom = new JSDOM(`<!doctype html><html><body>${html}</body></html>`, { runScripts: "dangerously" });
   bigtestGlobals.document = jsdom.window.document;
+  return jsdom.window;
 }
 
 beforeEach(() => {
